Sousou no freirein
Sousou no Freirein

A fantasy show where it shows the perspective of an elf and she takes care of a girl named fran and a boy named stark. A group of awkward adventures going on to find heaven and collecting weird magic book. It has a very amazing animation and battle. It is a very successful anime that is still on going.(1)